feat: gates and policies

+++ app/Http/Controllers/Api/AuthController.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\LoginRequest;
 use App\Http\Requests\RegisterUserRequest;
 use App\Models\User;
+use App\Models\UserToken;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
@@ -16,39 +17,17 @@ use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 use App\Policies\UserPolicy;
 
-use function PHPUnit\Framework\isEmpty;
-use function PHPUnit\Framework\isNull;
-
 class AuthController extends Controller
 {
     public function store(LoginRequest $request, AuthServices $authService)
     {
 
-        // return response()->json(request()->query());
-
-        // if (Gate::allows('login', $authService)) {
-        //     return response()->json(['error' => 'Already authenticated.'], 403);
-        // }
-        // $user = $authService->getUser();
-        // if (!$user) return response()->noContent(401);
-        // return response()->json($user);
-        // return response()->json($authService->getUser());
-
-
-        // if (Gate::denies('login', $authService->getUser())) {
-        //     return response()->json(['error' => 'Already authenticated.'], 403);
-        // }
-
-        // if ($authService->getUser())  $user = $authService->getUser();
-        // return  response()->json($authService->getUser() ? false : true);
-
-        // if (Gate::allows('create-token')) {
-        //     return response()->json(['error' => 'Already authenticated.'], 403);
-        // }
+        if (!Gate::denies('verified-user')) {
+            return response()->json(['error' => 'Not verified.'], 403);
+        }
 
 
 
-        // return response()->json('ok');
         $credentials = (['email' => $request->email, 'password' => $request->password]);
         if (auth()->guard('token')->attempt($credentials)) {
             $authService->createAccessToken();
@@ -61,6 +40,10 @@ class AuthController extends Controller
 
     public function destroy(Request $request, AuthServices $authService)
     {
+        if (!Gate::denies('owner-user')) {
+            return response()->json(['error' => 'Not an owner.'], 403);
+        }
+
         $authService->destroyAccessToken($request);
         return response()->noContent();
     }
@@ -72,11 +55,11 @@ class AuthController extends Controller
         return response()->json($registeredUser);
     }
 
-    public function verify(Request $request, TokenUserProvider $userProvider,  AuthServices $authService)
+    public function verify(Request $request, AuthServices $authService)
     {
+        $user = UserToken::where('access_token', $authService->getRequestToken())->first()->user;
+        $user->update(['is_verified' => $request->value]);
 
-        $user = auth()->guard('token')->user();
-        User::where('email', $user->email)->update(['is_verified' => $request->value]);
         return response()->noContent();
     }
